# Experiment with inlining dispatch VSDs