@benaadams 2022-11-26
Seems most effective for Bimorphic?