@jakobbotsch 2022-11-26
> Seems most effective for Bimorphic?

Yes, since the two call sites are both perfectly predictable by the CPU.