@jakobbotsch 2022-11-27
Need to spend some time looking at the failures before I can try this out on some larger benchmarks. Will try to find some time to do that soon.